Refactor handling of whitespace into helper
This should help create further shortcode and remain flexible with the markdown output
