+++ eleventy/shortcodes/grid.js
@@ -1,6 +1,6 @@
-import { outdent } from 'outdent'
+import { blockShortcode } from './util.js'
 
-export function grid(content, options = {}) {
+export const grid = blockShortcode((content, options = {}) => {
   const defaultOptions = {
     classes: undefined,
 
@@ -23,33 +23,5 @@ export function grid(content, options = {}) {
     cssProperties.push(`--app-grid-min-width: ${options.minWidth}`)
   }
 
-  // For content to be formatted properly inside the `<div>`
-  // we need two things to happen for markdown to be processed properly:
-  // 1. Removing any indentation, so that markdown blocks are properly at the start of the line
-  // 2. Adding new lines at the start and end, unless it's an HTML element
-  const formattedContent = ensureLeadingTrailingNewLines(
-    outdent.string(content)
-  )
-
-  return `\n\n<div class="app-grid${options.classes ? ` ${options.classes}` : ''}" style="${cssProperties.join(';')}">${formattedContent}</div>\n\n`
-}
-
-function ensureLeadingTrailingNewLines(content) {
-  content = content.trim()
-  if (!startsWithHTML(content)) {
-    content = `\n\n${content}`
-  }
-  if (!endsWithHTML(content)) {
-    content = `${content}\n\n`
-  }
-
-  return content
-}
-
-function startsWithHTML(string) {
-  return string.match(/^<\[\w][\w-]?>/)
-}
-
-function endsWithHTML(string) {
-  return string.match(/<\/[\w]+[\w-]?>$/)
-}
+  return `<div class="app-grid${options.classes ? ` ${options.classes}` : ''}" style="${cssProperties.join(';')}">${content}</div>`
+})
